# Rework SVMClassifier to improve performance