--- a/onnxruntime/core/providers/cpu/ml/ml_common.h
+++ b/onnxruntime/core/providers/cpu/ml/ml_common.h
@@ -185,16 +185,15 @@ static inline float ErfInv(float x) {
 }
 
 //https://www.csie.ntu.edu.tw/~cjlin/papers/svmprob/svmprob.pdf
-static inline void multiclass_probability(int64_t classcount, const std::vector<float>& r, std::vector<float>& p) {
+static inline void multiclass_probability(int64_t classcount,
+                                          const gsl::span<const float>& r,
+                                          const gsl::span<float>& p) {
   int64_t sized2 = classcount * classcount;
   std::vector<float> Q;
   std::vector<float> Qp;
-  for (int64_t k = 0; k < sized2; k++) {
-    Q.push_back(0);
-  }
-  for (int64_t k = 0; k < classcount; k++) {
-    Qp.push_back(0);
-  }
+  Q.assign(sized2, 0.f);
+  Qp.assign(classcount, 0.f);
+
   float eps = 0.005f / static_cast<float>(classcount);
   for (int64_t i = 0; i < classcount; i++) {
     p[i] = 1.0f / static_cast<float>(classcount);  // Valid if k = 1
@@ -207,6 +206,7 @@ static inline void multiclass_probability(int64_t classcount, const std::vector<
       Q[i * classcount + j] = -r[j * classcount + i] * r[i * classcount + j];
     }
   }
+
   for (int64_t loop = 0; loop < 100; loop++) {
     // stopping condition, recalculate QP,pQP for numerical accuracy
     float pQp = 0;
@@ -217,6 +217,7 @@ static inline void multiclass_probability(int64_t classcount, const std::vector<
       }
       pQp += p[i] * Qp[i];
     }
+
     float max_error = 0;
     for (int64_t i = 0; i < classcount; i++) {
       float error = std::fabs(Qp[i] - pQp);
@@ -224,7 +225,9 @@ static inline void multiclass_probability(int64_t classcount, const std::vector<
         max_error = error;
       }
     }
-    if (max_error < eps) break;
+
+    if (max_error < eps)
+      break;
 
     for (int64_t i = 0; i < classcount; i++) {
       float diff = (-Qp[i] + pQp) / Q[i * classcount + i];
@@ -394,7 +397,8 @@ static void write_scores(std::vector<T>& scores, POST_EVAL_TRANSFORM post_transf
 //       was seen from testing with the arbitrary values of 1000 scores per threads.
 template <typename T>
 void batched_update_scores_inplace(gsl::span<T> scores, int64_t num_batches_in, int64_t batch_size,
-                                   POST_EVAL_TRANSFORM post_transform, int add_second_class,
+                                   POST_EVAL_TRANSFORM post_transform,
+                                   int add_second_class, bool have_space_for_second_class,
                                    concurrency::ThreadPool* threadpool) {
   if (batch_size < 1)
     return;
@@ -484,12 +488,22 @@ void batched_update_scores_inplace(gsl::span<T> scores, int64_t num_batches_in, 
           ORT_THROW("Unexpected value for 'add_second_class' of ", add_second_class);
       }
 
-      const float* cur_in = s_end;
-      float* cur_out = &*scores.end();
-      while (cur_in > s) {
-        --cur_in;
-        cur_out -= 2;
-        update_scores(*cur_in, cur_out);
+      if (have_space_for_second_class) {
+        // forward iteration as there's a gap between each score to write into
+        float* cur_score = scores.data();
+        for (int i = 0; i < num_batches; ++i) {
+          update_scores(*cur_score, cur_score);
+          cur_score += 2;
+        }
+      } else {
+        // reverse iteration as the scores are packed together and each score needs to be expanded to two
+        const float* cur_in = s_end;
+        float* cur_out = &*scores.end();
+        while (cur_in > s) {
+          --cur_in;
+          cur_out -= 2;
+          update_scores(*cur_in, cur_out);
+        }
       }
     }
   }