--- a/onnxruntime/core/providers/cpu/ml/svmregressor.h
+++ b/onnxruntime/core/providers/cpu/ml/svmregressor.h
@@ -13,11 +13,10 @@ namespace onnxruntime {
 namespace ml {
 
 template <typename T>
-class SVMRegressor final : public OpKernel, private SVMCommon<T> {
-  using SVMCommon<T>::batched_kernel_dot;
-  using SVMCommon<T>::kernel_dot;
-  using SVMCommon<T>::set_kernel_type;
-  using SVMCommon<T>::get_kernel_type;
+class SVMRegressor final : public OpKernel, private SVMCommon {
+  using SVMCommon::batched_kernel_dot;
+  using SVMCommon::set_kernel_type;
+  using SVMCommon::get_kernel_type;
 
  public:
   SVMRegressor(const OpKernelInfo& info);