**Description**: 
 - use GEMM for initial scoring
 - minimize data allocations and copies
 - parallelize the second half of the scoring for larger batches
   - arbitrary minimum batch size of 512 to start with

**Motivation and Context**
Close gap with SKL performance


**Performance improvement**
Local testing using 2 tests from the SKL benchmarks. Compared current master with SKL, and new code with SKL. Numbers are 'average SKL latency'/'average ORT latency' for 5 seconds of execution for the different combinations. 

svm/NuSVC/bench_NuSVC_prob_b_cl_probabilityTrue.py  


  |   | SKL/ORT | SKL/ORT
-- | -- | -- | --
  |   | MASTER | This PR
Batch size=1 | Features=4 | 2.20435 | 2.4234
Batch size=1 | Features=20 | 2.11115 | 2.28475
Batch size=10 | Features=4 | 1.5964 | 2.17515
Batch size=10 | Features=20 | 1.5559 | 2.0326
Batch size=100 | Features=4 | 1.09125 | 1.8263
Batch size=100 | Features=20 | 1.11215 | 1.6694
Batch size=1000 | Features=4 | 1.01255 | 1.9508
Batch size=1000 | Features=20 | 1.02555 | 1.66325
Batch size=10000 | Features=4 | 1.004 | 1.98565
Batch size=10000 | Features=20 | 0.96995 | 1.78045
Batch size=100000 | Features=4 | 1.26025 | 2.3442
Batch size=100000 | Features=20 | 1.00765 | 1.9366


svm/SVC/bench_SVC_poly_b_cl_probabilityTrue_kernelpoly.py


  |   | MASTER | This PR
-- | -- | -- | --
Batch size=1 | Features=4 | 2.153 | 2.572
Batch size=1 | Features=20 | 2.0552 | 2.4353
Batch size=10 | Features=4 | 1.40085 | 2.59095
Batch size=10 | Features=20 | 1.1492 | 2.64675
Batch size=100 | Features=4 | 0.9558 | 2.65655
Batch size=100 | Features=20 | 0.63875 | 1.91535
Batch size=1000 | Features=4 | 0.60285 | 4.44615
Batch size=1000 | Features=20 | 0.54305 | 7.7014
Batch size=10000 | Features=4 | 0.58395 | 4.7105
Batch size=10000 | Features=20 | 0.5611 | 5.90245
Batch size=100000 | Features=4 | 0.5816 | 7.16635
Batch size=100000 | Features=20 | 0.5428 | 7.73235